Move isFeaturedEnabled to shared
